delete, log

+++ Board/src/main/java/com/board/controller/BoardController.java
@@ -10,11 +10,13 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.board.constant.Method;
 import com.board.domain.BoardDTO;
 import com.board.service.BoardService;
+import com.board.util.UiUtils;
 
 @Controller
-public class BoardController {
+public class BoardController extends UiUtils{
 
 	@Autowired
 	private BoardService boardService;
@@ -51,23 +53,27 @@ public class BoardController {
 	
 	
 	@PostMapping(value = "/board/register.do")
-	public String registerBoard(final BoardDTO params) {
+	public String registerBoard(final BoardDTO params, Model model) {
 	
 		try {
 			boolean isRegisterd = boardService.registerBoard(params);
 			if(isRegisterd == false) {
-				//등록 실패
+				return showMessageWithRedirect("게시글 등록에 실패하였습니다.", "/board/list.do", Method.GET, null, model);
 			}
 		} catch (DataAccessException e) {
-			// 데이터베이서 처리 문제
+			return showMessageWithRedirect("데이터베이스 처리 과정에 문제가 생겼습니다.", "/board/list.do", Method.GET, null, model);
+
 		} catch (Exception e) { 
-			// 시스템 문제
+			return showMessageWithRedirect("시스템 문제가 발생했습니다.", "/board/list.do", Method.GET, null, model);
+
 		}
 		
-		return "redirect:/board/list.do";
+		//return "redirect:/board/list.do";
+		return showMessageWithRedirect("게시글 등록이 완료되었습니다.", "/board/list.do", Method.GET, null, model);
 		
 	}
 	
+
 	@GetMapping(value = "/board/list.do")
 	public String openBoardList(Model model) {
 		
@@ -81,20 +87,46 @@ public class BoardController {
 	@GetMapping(value = "/board/view.do")
 	public String boardDetail(@RequestParam(value = "idx", required = false) Long idx, Model model) {
 		
+		if(idx == null) {
+			return "redirect:/board/list.do";
+		}
+		
 		BoardDTO board = boardService.getBoardDetail(idx);
+		if(board == null || "Y".equals(board.getDeleteYn())) {
+			return  "redirect:/board/list.do";
+		}
 		model.addAttribute("board", board);
 		
 		return "board/view";
 	}
 	
 	
-	@GetMapping(value = "/board/delete.do")
+	@PostMapping(value = "/board/delete.do")
 	public String boardDelete(@RequestParam(value = "idx", required = false) Long idx, Model model) {
 		
-		boolean board = boardService.deleteBoard(idx);
-		model.addAttribute("board", board);
+		//System.out.println("/board/delete.do : idx = "+idx);
 		
-		return "board/list";
+		if(idx == null) {
+			return showMessageWithRedirect("올바르지 않은 접근입니다.", "/board/list.do", Method.GET, null, model);
+			//return "redirect:/board/list.do";
+		}
+		
+		try {
+			boolean deleted = boardService.deleteBoard(idx);
+			if (deleted == false) {
+				return showMessageWithRedirect("게시글 삭제에 실패하였습니다.", "/board/list.do", Method.GET, null, model);
+			}
+		} catch (DataAccessException e) {
+			return showMessageWithRedirect("데이터베이스 처리 과정에 문제가 생겼습니다.", "/board/list.do", Method.GET, null, model);
+
+		} catch (Exception e) { 
+			return showMessageWithRedirect("시스템 문제가 발생했습니다.", "/board/list.do", Method.GET, null, model);
+
+		}
+		//model.addAttribute("board", board);
+		
+		return showMessageWithRedirect("게시글 삭제가 완료되었습니다.", "/board/list.do", Method.GET, null, model);
+		//return "redirect:/board/list.do";
 	}
